PuzzleCore #16
- Deleted comments, refactored the code

+++ Assets/Puzzles/Pipes/Scripts/PipesLevelData.cs
@@ -11,7 +11,6 @@ namespace Pipes
         public List<int> Data = new List<int>();
         private void OnValidate()
         {
-            // Если значения Row или Col изменились, обновить размер списка Data
             int newSize = Row * Col;
             if (Data.Count != newSize)
             {
@@ -19,10 +18,8 @@ namespace Pipes
                 {
                     Data.Add(0);
                 }
-                if (Data.Count > newSize)
-                {
-                    Data.RemoveRange(newSize, Data.Count - newSize);
-                }
+
+                if (Data.Count > newSize) Data.RemoveRange(newSize, Data.Count - newSize);
             }
         }
     }
